feat: Se creó la api para crear usuarios con sus datos personales y negocio

+++ FacontaryFunctions/Dto/Negocio/NegocioInput.cs
@@ -6,6 +6,6 @@ namespace FacontaryFunctions.Dto.Negocio
 {
     class NegocioInput
     {
-        public int IdNegocio { get; set; }
+        public long IdNegocio { get; set; }
     }
 }
